+++ provider/pkg/await/await.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2022, Pulumi Corporation.
+// Copyright 2016-2023, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"regexp"
 	"strings"
 
 	"github.com/pulumi/pulumi-kubernetes/provider/v3/pkg/clients"
@@ -140,6 +141,11 @@ func skipRetry(gvk schema.GroupVersionKind, k8sVersion *cluster.ServerVersion, e
 }
 
 const ssaConflictDocLink = "https://www.pulumi.com/registry/packages/kubernetes/how-to-guides/managing-resources-with-server-side-apply/#handle-field-conflicts-on-existing-resources"
+
+// Resources created with Client-side Apply (CSA) will assign a field manager with an operation of type "Update" that
+// conflicts with Server-side Apply (SSA) field managers that use type "Apply". This regex is used to check for known
+// CSA field manager names to determine if the apply operation should be retried.
+var csaConflictRegex = regexp.MustCompile(`conflict with "(pulumi-resource-kubernetes|pulumi-resource-kubernetes.exe|pulumi-kubernetes)"`)
 
 // Creation (as the usage, `await.Creation`, implies) will block until one of the following is true:
 // (1) the Kubernetes resource is reported to be initialized; (2) the initialization timeout has
@@ -438,13 +444,29 @@ func Update(c UpdateConfig) (*unstructured.Unstructured, error) {
 			// Issue patch request.
 			// NOTE: We can use the same client because if the `kind` changes, this will cause
 			// a replace (i.e., destroy and create).
-			currentOutputs, err = client.Patch(c.Context, c.Inputs.GetName(), types.ApplyPatchType, objYAML, options)
-			if err != nil {
-				if errors.IsConflict(err) {
-					err = fmt.Errorf("Server-Side Apply field conflict detected. see %s for troubleshooting help\n: %w",
-						ssaConflictDocLink, err)
+			maybeRetry := true
+			for {
+				currentOutputs, err = client.Patch(c.Context, c.Inputs.GetName(), types.ApplyPatchType, objYAML, options)
+				if err != nil {
+					if errors.IsConflict(err) {
+						if maybeRetry {
+							// If the patch failed, use heuristics to determine if the resource was created using
+							// Client-side Apply. If so, retry once with the force patch option.
+							if csaConflictRegex.MatchString(err.Error()) &&
+								metadata.GetLabel(c.Inputs, metadata.LabelManagedBy) == "pulumi" {
+								force = true
+								options.Force = &force
+								maybeRetry = false // Only retry once
+								continue
+							}
+						}
+
+						err = fmt.Errorf("Server-Side Apply field conflict detected. see %s for troubleshooting help\n: %w",
+							ssaConflictDocLink, err)
+					}
+					return nil, err
 				}
-				return nil, err
+				break
 			}
 
 		} else {